Change type of calorie

+++ HealthyLife.Data/Entities/Food/Ingredient.cs
@@ -10,7 +10,7 @@
 
         // Nutritions
         // Everything is measured in 100 g
-        public int Calory { get; set; } // kcal
+        public double Calory { get; set; } // kcal
         public double Carbohydrate { get; set; }
         public double Protein { get; set; }
         public double Fat { get; set; }
