+++ HealthyLife.Models/Food/IngredientModel.cs
@@ -12,7 +12,7 @@ namespace HealthyLife.Models.Food
 
         // Nutritions
         // Everything is measured in 100 g
-        public int Calory { get; set; } // kcal
+        public double Calory { get; set; } // kcal
         public double Carbohydrate { get; set; }
         public double Protein { get; set; }
         public double Fat { get; set; }
